Fixed time gap bug; changing music to node generator (under same fixedupdate)
we used Time.fixed - timer >= timePerTempo to generate node, and we reset the timer in the if function. All the code in between will cause delay of the next detection, which result in a overall delay for node generation.

Now we use Time.fixedTime >= initialTime + timePerTempo * counter, which means all the moment is a fixed time, not a calculated time, so it's much more accurate than before.